backend/requirements.production.txt
# Production Requirements for NCLEX Website
# Install with: pip install -r requirements.production.txt

# Django Core
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==24.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10
dj-database-url==3.0.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# File Upload & Media
Pillow==11.3.0
python-magic==0.4.27
opencv-python==4.12.0.88
moviepy==2.2.1

# Payment Processing
requests==2.32.4

# Email & Notifications
django-anymail==10.2

# Caching & Performance
redis>=4.5.2,<5.0.0
django-redis==5.4.0
celery[redis]==5.3.4
django-celery-beat==2.8.1

# Environment & Configuration
python-dotenv==1.1.1
django-environ==0.12.0

# Production Server
gunicorn==23.0.0
whitenoise==6.9.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Cloudinary
cloudinary==1.44.1
django-cloudinary-storage==0.3.0

# Additional Dependencies
PyJWT==2.10.1
python-decouple==3.8

# Development & Testing (Optional for production)
# pytest==7.4.3
# pytest-django==4.7.0
# factory-boy==3.3.0
